Switch classproperty implementation.
My implementation doesn't make much sense now that I know that a
`property` can be overriden by a class attribute in a subclass. I find
this behavior odd though, because you cannot depend on a `property` to
behave the same as an instance attribute, which I thought was the whole
point.